Remove unused SVG assets
This commit removes several SVG files that were no longer in use within the project. The deleted files include various ellipse and line assets, as well as a home indicator and right-side SVG. Additionally, the commit removes a Vuesax linear arrow left icon and a specific image file related to a right-side element.